moved config data logic to trait
